polling package refactor
