--- src/main/java/com/configcat/ManualPollingMode.java
@@ -1,11 +1,0 @@
-package com.configcat;
-
-/**
- * The manual polling mode configuration.
- */
-class ManualPollingMode extends PollingMode {
-    @Override
-    String getPollingIdentifier() {
-        return "m";
-    }
-}
