+++ src/main/java/com/configcat/polling/ManualPollingMode.java
@@ -1,0 +1,11 @@
+package com.configcat.polling;
+
+/**
+ * The manual polling mode configuration.
+ */
+public class ManualPollingMode extends PollingMode {
+    @Override
+    public String getPollingIdentifier() {
+        return "m";
+    }
+}
