+++ src/test/java/com/configcat/ConfigCatClientIntegrationTest.java
@@ -2,6 +2,7 @@ package com.configcat;
 
 import com.configcat.evaluation.EvaluationDetails;
 import com.configcat.log.LogLevel;
+import com.configcat.polling.PollingModes;
 import okhttp3.OkHttpClient;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
