+++ src/test/java/com/configcat/ConfigFetcherTest.java
@@ -2,6 +2,7 @@ package com.configcat;
 
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
+import com.configcat.polling.PollingModes;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import okhttp3.OkHttpClient;
