+++ src/test/java/com/configcat/LocalTests.java
@@ -2,6 +2,7 @@ package com.configcat;
 
 import com.configcat.override.OverrideBehaviour;
 import com.configcat.override.OverrideDataSourceBuilder;
+import com.configcat.polling.PollingModes;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import org.junit.jupiter.api.Test;
